Correct and clarify HTML inputs
E.g. `<div/>` is not valid in HTML documents, only in XHTML documents (which we’re not using here). It still appeared to work because the HTML parser is very forgiving, but this patch clarifies the intended result by explicitly including closing tags where applicable (even in cases where they are not strictly necessary for validity, like `<p>`).